refactor(runtime): extract executeStream into engine/execute-stream (W26-A step 3)
Final W26-A extraction. The executeStream method body (~145 LOC of Queue +
EventBus + FiberRef daemon plumbing) moves into engine/execute-stream.ts as a
makeExecuteStream({config, execute}) factory. Inline as-any casts replaced with
typed structural assertions during the move.

- Behavior preserved: runtime 811/0/1 + replay 24/0 + build 38/38.
- LOC: 1522 → 1378 (-144).
- Cumulative W26-A: 1656 → 1378 (-278 / -17%) — comfortably under #76's
  ≤1500 LOC threshold for execution-engine.ts.

Partial: #76
