+++ packages/runtime/src/execution-engine.ts
@@ -72,6 +72,7 @@ import { runIterationGuards } from "./engine/phases/agent-loop/iteration-guards.
 import { runBootstrapSkillPostprocess } from "./engine/bootstrap/skill-postprocess.js";
 import { runPreLoopDispatch } from "./engine/phases/agent-loop/setup/pre-loop-dispatch.js";
 import { captureFinalSnapshot } from "./engine/finalize/snapshot-final.js";
+import { makeExecuteStream } from "./engine/execute-stream.js";
 import { synthesizeAndStoreDebrief } from "./engine/finalize/debrief-synthesis.js";
 import { emitTelemetryRunReport } from "./engine/finalize/telemetry-emit.js";
 import { runLocalLearning } from "./engine/finalize/local-learning.js";
@@ -1371,152 +1372,7 @@ export const ExecutionEngineLive = (config: ReactiveAgentsConfig) =>
             yield* Ref.update(cancelledTasks, (s) => new Set(s).add(taskId));
           }),
 
-        executeStream: (task, options) =>
-          Effect.gen(function* () {
-            const queue = yield* Queue.unbounded<AgentStreamEvent>();
-            const density = options?.density ?? config.streamDensity ?? "tokens";
-            const startMs = Date.now();
-
-            // Acquire EventBus optionally
-            const ebOpt = yield* Effect.serviceOption(EventBus).pipe(
-              Effect.catchAll(() => Effect.succeed({ _tag: "None" as const })),
-            );
-            const eb: EbLike | null = ebOpt._tag === "Some" ? ebOpt.value : null;
-
-            // Fire AgentStreamStarted
-            if (eb) {
-              yield* eb.publish({
-                _tag: "AgentStreamStarted",
-                taskId: String(task.id),
-                agentId: config.agentId,
-                density,
-                timestamp: startMs,
-              } as AgentEvent).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4277", tag: errorTag(err) })));
-            }
-
-            // Subscribe to ReasoningIterationProgress events — push them as IterationProgress stream events
-            if (eb) {
-              yield* eb.on("ReasoningIterationProgress", (event) =>
-                Effect.gen(function* () {
-                  const eventTaskId = String((event as { taskId?: string }).taskId ?? "");
-                  if (eventTaskId !== String(task.id)) {
-                    return;
-                  }
-                  yield* Queue.offer(queue, {
-                    _tag: "IterationProgress",
-                    iteration: event.iteration,
-                    maxIterations: event.maxIterations,
-                    toolsCalledThisStep: event.toolsThisStep,
-                    status: `iteration ${event.iteration}/${event.maxIterations}`,
-                  } as AgentStreamEvent).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4294", tag: errorTag(err) })));
-                }),
-              ).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4296", tag: errorTag(err) })));
-            }
-
-            // Full density: emit live thought text for each reasoning step.
-            if (eb && density === "full") {
-              yield* eb.on("ReasoningStepCompleted", (event) =>
-                Effect.gen(function* () {
-                  const eventTaskId = String((event as { taskId?: string }).taskId ?? "");
-                  if (eventTaskId !== String(task.id)) {
-                    return;
-                  }
-                  const thought =
-                    typeof (event as { thought?: unknown }).thought === "string"
-                      ? ((event as { thought?: string }).thought ?? "").trim()
-                      : "";
-                  if (thought.length === 0) {
-                    return;
-                  }
-                  const iteration =
-                    typeof (event as { step?: unknown }).step === "number"
-                      ? ((event as { step?: number }).step ?? 0)
-                      : 0;
-                  yield* Queue.offer(queue, {
-                    _tag: "ThoughtEmitted",
-                    content: thought,
-                    iteration,
-                  } as AgentStreamEvent).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4322", tag: errorTag(err) })));
-                }),
-              ).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4324", tag: errorTag(err) })));
-            }
-
-            // Set the streaming callback inside the daemon so the FiberRef is
-            // available to the reasoning kernel. forkDaemon creates a root fiber
-            // that does NOT inherit FiberRef values from Effect.locally — the only
-            // reliable way is FiberRef.set as the first step inside the fork.
-            const streamCallback = (text: string) =>
-              Queue.offer(queue, { _tag: "TextDelta", text }).pipe(
-                Effect.map(() => {}),
-              );
-
-            yield* FiberRef.set(StreamingTextCallback, streamCallback).pipe(
-              Effect.andThen(FiberRef.set(RunControllerRef, options?.runController ?? null)),
-              Effect.andThen(execute(task)),
-              Effect.tap((taskResult) => {
-                const debriefToolsUsed = (taskResult as any).debrief?.toolsUsed as Array<{ name: string; calls: number; successRate: number }> | undefined;
-                const toolSummary = debriefToolsUsed && debriefToolsUsed.length > 0
-                  ? debriefToolsUsed.map((t) => ({ name: t.name, calls: t.calls, avgMs: 0 }))
-                  : [];
-                const completedEvent: AgentStreamEvent = {
-                  _tag: "StreamCompleted",
-                  output: String((taskResult as any).output ?? ""),
-                  metadata: (taskResult as any).metadata ?? {},
-                  taskId: String(task.id),
-                  agentId: String(task.agentId),
-                  ...(toolSummary.length > 0 ? { toolSummary } : {}),
-                };
-                const offer = Queue.offer(queue, completedEvent);
-                if (!eb) return offer;
-                return offer.pipe(
-                  Effect.tap(() =>
-                    eb.publish({
-                      _tag: "AgentStreamCompleted",
-                      taskId: String(task.id),
-                      agentId: config.agentId,
-                      success: true,
-                      durationMs: Date.now() - startMs,
-                    } as AgentEvent).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4361", tag: errorTag(err) }))),
-                  ),
-                );
-              }),
-              Effect.catchAll((err: unknown) => {
-                const cause =
-                  typeof err === "object" && err !== null && "message" in err
-                    ? String((err as any).message)
-                    : String(err);
-                const errorEvent: AgentStreamEvent = { _tag: "StreamError", cause };
-                const offer = Queue.offer(queue, errorEvent);
-                if (!eb) return offer;
-                return offer.pipe(
-                  Effect.tap(() =>
-                    eb.publish({
-                      _tag: "AgentStreamCompleted",
-                      taskId: String(task.id),
-                      agentId: config.agentId,
-                      success: false,
-                      durationMs: Date.now() - startMs,
-                    } as AgentEvent).pipe(Effect.catchAll((err) => emitErrorSwallowed({ site: "runtime/src/execution-engine.ts:4381", tag: errorTag(err) }))),
-                  ),
-                );
-              }),
-              Effect.forkDaemon,
-            );
-
-            // Stream reads from queue, stops after terminal event.
-            return EStream.unfoldEffect(false as boolean, (done) => {
-              if (done) return Effect.succeed(Option.none());
-              return Queue.take(queue).pipe(
-                Effect.map((event) => {
-                  const isTerminal =
-                    event._tag === "StreamCompleted" ||
-                    event._tag === "StreamError" ||
-                    event._tag === "StreamCancelled";
-                  return Option.some([event, isTerminal] as const);
-                }),
-              );
-            });
-          }),
+        executeStream: makeExecuteStream({ config, execute }),
       };
     }),
   );
